Set heap and permgen sizes to maximum values
Set -Xms to the same value as -Xms and -XX:PermSize to the same value
as -XX:MaxPermSize. This ensures that the committed memory for the
heap and the permanent generation does not increase during the
lifetime of the application. This should reduce the possibility of
warden killing the JVM process once the application is running.

[#59731874]